added select object by name feature
on creation of an object the user will be prompt for an object name,
object names are unique identifiers of the objects in the scene.
on 'z' press the option select active model will be activated in the
command line. the user will be presented with the object names and will
be asked to enter a name, the object who's name was entered will be the
active object.
objects are now saved in a map with map<string,Model*>
signature.